+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## [Unreleased]
+### Added
+- [#28] (https://github.com/BixData/stuart-ml/issues/28) Train a K-means model
+
 ### Fixed
-- [#23](https://github.com/BixData/stuart-ml/issues/23) KMeans cluster centers load in a non-deterministic sort order due to missing sort
+- [#23](https://github.com/BixData/stuart-ml/issues/23) K-means cluster centers load in a non-deterministic sort order due to missing sort
 - [#25](https://github.com/BixData/stuart-ml/issues/25) VectorWithNorm initializer fails to compute norm when not explicitly provided
 
 ## [0.1.8] - 2018-10-14